Add meta image metadata
For the version "$DOCKER_TAG" is being used instead of "v$JIRA_VERSION",
following instructions from http://label-schema.org/rc1/:

> The version MAY match a label or tag in the source code repository.
You should make sure that only images that exactly reflect a version of
code should have that version label. If Julia finds a version 0.7.1 in a
repository she SHOULD be able to infer this matches version 0.7.1 of the
associated code (and in particular, not 0.7.1 plus some later commits).

Would I had used the version in the $JIRA_VERSION, changes made to the
Dockerfile without changing the jira version would have generated a new
image for a metadata version that wouldn't match the source code tag.

With this method two types of metadata versions are generated:
- `vX.X.X`: that PROPERLY match a git tag.
- `latest` and `master`: that do not match a git tag (they don't need to).